LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/stylevalue-serialization/cssTransformValue.tentative-expected.txt

CONSOLE MESSAGE: ReferenceError: Can't find variable: CSSTranslate

Harness Error (FAIL), message = ReferenceError: Can't find variable: CSSTranslate
